mcp-servers/debugger/requirements.txt
mcp>=0.1.0
boto3>=1.28.0
python-dotenv>=1.0.0
pytest>=8.4.1
black>=25.1.0